deprecated lsst-backup.py and rename dask version as replacement
